Fixed datatype for state

+++ pennylane/qchem/structure.py
@@ -366,7 +366,7 @@ def hf_state(electrons, orbitals, basis="occupation_number"):
 
     if basis == "parity":
         pi_matrix = np.tril(np.ones((orbitals, orbitals)))
-        return (np.matmul(pi_matrix, state) % 2).astype("int32")
+        return (np.matmul(pi_matrix, state) % 2).astype(int)
 
     if basis == "bravyi_kitaev":
         beta_matrix = _beta_matrix(orbitals)
